THe Fisher's method for meta-analysis
This R script contains one of the methods used for the meta-analysis in this research project. 
However, the datasets was very heterogeneous. Therefore we proceeded with the z-transformation method for the final report instead of fisher's method in order to account for the heterogeneity of the data we included in our meta-analysis.