Combat to handle batch affect
One of the methods used to attempt to handle batch effect is the Combat method. 
However, after normalising the data, performing DEG with limma, and comparing the results with other meta-analysis methods used on the same dataset,.
We noticed from the PCA that the Combat method was not very useful in handling batch effects. Therefore, it was omitted from the final results.